Moved the server-only code in the code only loaded on the server
